core-aam/arianotify/aria-notify-iframe-permissions-policy.html

<!DOCTYPE html>
<html>
<head>
  <title>ARIA Notify API Test: iframe Permissions Policy (Non-Manual)</title>
  <link rel="help" href="https://github.com/w3c/aria/pull/2577">
  <link rel="author" title="Jacques Newman" href="mailto:janewman@microsoft.com">
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
  <script>
  "use strict";
  const iframe_src = 'core-aam/arianotify/aria-notify-iframe-permissions-policy.html/resources/aria-notify-iframe-test.html';

  function test_aria_notify_in_iframe(test_name, iframe_allow_attribute, expect_allowed) {
    async_test(t => {
      const iframe = document.createElement('iframe');
      if (iframe_allow_attribute !== null) {
        iframe.allow = iframe_allow_attribute;
      }
      iframe.src = iframe_src;

      window.addEventListener('message', t.step_func(function(e) {
        if (e.source !== iframe.contentWindow) return;

        assert_true(e.data.supportsAriaNotify, 'ariaNotify API should be supported');

        if (expect_allowed) {
          assert_true(e.data.allowsFeature, `Expected aria-notify to be allowed by permissions policy`);
        } else {
          assert_false(e.data.allowsFeature, `Expected aria-notify to be disabled by permissions policy`);
        }

        iframe.remove();
        t.done();
      }));

      document.body.appendChild(iframe);
    }, test_name);
  }

  // Test default iframe behavior (should allow aria-notify).
  test_aria_notify_in_iframe(
    'aria-notify is enabled in iframe with default permissions policy',
    null,
    true
  );

  // Test denied iframe.
  test_aria_notify_in_iframe(
    'aria-notify is disabled in iframe with allow="aria-notify \'none\'" attribute',
    'aria-notify \'none\'',
    false
  );
  </script>
</body>
</html>
